+++ test/python/visualization/test_circuit_matplotlib_drawer.py
@@ -62,56 +62,6 @@ class TestMatplotlibDrawer(QiskitVisualizationTestCase):
 
         self.assertImagesAreEqual(filename, expected_filename)
 
-    @unittest.skipIf(not visualization.HAS_MATPLOTLIB, 'matplotlib not available.')
-    def test_plot_barriers(self):
-        """Test to see that plotting barriers works.
-        If it is set to False, no blank columns are introduced"""
-
-        # generate a circuit with barriers and other barrier like instructions in
-        q = QuantumRegister(2, 'q')
-        c = ClassicalRegister(2, 'c')
-        qc = QuantumCircuit(q, c)
-
-        # check for barriers
-        qc.h(q[0])
-        qc.barrier()
-
-        # check for other barrier like commands
-        qc.h(q[1])
-
-        # this import appears to be unused, but is actually needed to get snapshot instruction
-        import qiskit.extensions.simulator  # pylint: disable=unused-import
-        qc.snapshot('1')
-
-        # check the barriers plot properly when plot_barriers= True
-        filename = self._get_resource_path('visualization/references/current_matplotlib_ref.png')
-        visualization.circuit_drawer(qc, output='mpl', plot_barriers=True, filename=filename)
-        self.addCleanup(os.remove, filename)
-
-        ref_filename = self._get_resource_path(
-            'visualization/references/matplotlib_barriers_ref.png')
-        self.assertImagesAreEqual(filename, ref_filename)
-
-        # check that the barrier aren't plotted when plot_barriers = False
-        filename = self._get_resource_path('current_matplotlib_ref.png')
-        visualization.circuit_drawer(qc, output='mpl', plot_barriers=False, filename=filename)
-        self.addCleanup(os.remove, filename)
-
-        # generate the same circuit but without the barrier commands as this is what the
-        # circuit should look like when displayed with plot barriers false
-        q1 = QuantumRegister(2, 'q')
-        c1 = ClassicalRegister(2, 'c')
-        qc1 = QuantumCircuit(q1, c1)
-        qc1.h(q1[0])
-        qc1.h(q1[1])
-
-        no_barriers_filename = self._get_resource_path('current_no_barriers_matplotlib_ref.png')
-        visualization.circuit_drawer(qc1, output='mpl', justify='None',
-                                     filename=no_barriers_filename)
-        self.addCleanup(os.remove, no_barriers_filename)
-
-        self.assertImagesAreEqual(filename, no_barriers_filename)
-
     @unittest.skipIf(not visualization.HAS_MATPLOTLIB,
                      'matplotlib not available.')
     @unittest.skip('Unreliable across python version')
@@ -139,27 +89,3 @@ class TestMatplotlibDrawer(QiskitVisualizationTestCase):
             'visualization/references/%s_long_name_matplotlib.png' % os.name)
 
         self.assertImagesAreEqual(ref_filename, filename)
-
-    @unittest.skipIf(not visualization.HAS_MATPLOTLIB,
-                     'matplotlib not available.')
-    def test_conditional(self):
-        """Test that circuits with conditionals draw correctly
-        """
-        qr = QuantumRegister(2, 'q')
-        cr = ClassicalRegister(2, 'c')
-        circuit = QuantumCircuit(qr, cr)
-
-        # check gates are shifted over accordingly
-        circuit.h(qr)
-        circuit.measure(qr, cr)
-        circuit.h(qr[0]).c_if(cr, 2)
-
-        conditional_filename = self._get_resource_path('current_conditional_matplotlib_ref.png')
-        visualization.circuit_drawer(circuit, output='mpl',
-                                     filename=conditional_filename)
-        self.addCleanup(os.remove, conditional_filename)
-
-        ref_filename = self._get_resource_path(
-            'visualization/references/matplotlib_conditional_ref.png')
-
-        self.assertImagesAreEqual(ref_filename, conditional_filename, diff_tolerance=0.002)
